+++ packages/clients/src/scw/constants.ts
@@ -1,3 +1,0 @@
-export const version = 'v2.67.0'
-
-export const userAgent = `scaleway-sdk-js/${version}`