+++ packages_generated/lb/src/v1/validation-rules.gen.ts
@@ -1,0 +1,20 @@
+// This file was automatically generated. DO NOT EDIT.
+// If you have any remark or suggestion do not hesitate to open an issue.
+
+export const HealthCheckHttpConfig = {
+  hostHeader: {
+    pattern: /^[^\s]+$/,
+  },
+  uri: {
+    pattern: /^[^\s]+$/,
+  },
+}
+
+export const HealthCheckHttpsConfig = {
+  hostHeader: {
+    pattern: /^[^\s]+$/,
+  },
+  uri: {
+    pattern: /^[^\s]+$/,
+  },
+}