+++ packages_generated/product_catalog/src/v2alpha1/index.gen.ts
@@ -1,0 +1,33 @@
+// This file was automatically generated. DO NOT EDIT.
+// If you have any remark or suggestion do not hesitate to open an issue.
+export { PublicCatalogAPI } from './api.gen'
+export * from './marshalling.gen'
+export type {
+  ListPublicCatalogProductsRequestProductType,
+  ListPublicCatalogProductsResponse,
+  PublicCatalogApiListPublicCatalogProductsRequest,
+  PublicCatalogProduct,
+  PublicCatalogProductEnvironmentalImpactEstimation,
+  PublicCatalogProductLocality,
+  PublicCatalogProductPrice,
+  PublicCatalogProductProperties,
+  PublicCatalogProductPropertiesAppleSilicon,
+  PublicCatalogProductPropertiesBlockStorage,
+  PublicCatalogProductPropertiesDedibox,
+  PublicCatalogProductPropertiesElasticMetal,
+  PublicCatalogProductPropertiesHardware,
+  PublicCatalogProductPropertiesHardwareCPU,
+  PublicCatalogProductPropertiesHardwareCPUArch,
+  PublicCatalogProductPropertiesHardwareCPUPhysical,
+  PublicCatalogProductPropertiesHardwareCPUVirtual,
+  PublicCatalogProductPropertiesHardwareGPU,
+  PublicCatalogProductPropertiesHardwareNetwork,
+  PublicCatalogProductPropertiesHardwareRAM,
+  PublicCatalogProductPropertiesHardwareStorage,
+  PublicCatalogProductPropertiesInstance,
+  PublicCatalogProductPropertiesObjectStorage,
+  PublicCatalogProductStatus,
+  PublicCatalogProductUnitOfMeasure,
+  PublicCatalogProductUnitOfMeasureCountableUnit,
+} from './types.gen'
+export * as ValidationRules from './validation-rules.gen'