+++ packages/configuration-loader/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 1.0.9 (2025-05-26)
+
+**Note:** Version bump only for package @scaleway/configuration-loader
+
+## 1.0.8 (2025-05-13)
+
+**Note:** Version bump only for package @scaleway/configuration-loader
+
+## [1.0.7](https://github.com/scaleway/scaleway-sdk-js/compare/@scaleway/configuration-loader@1.0.6...@scaleway/configuration-loader@1.0.7) (2025-03-31)
+
+**Note:** Version bump only for package @scaleway/configuration-loader
+
+## 1.0.6 (2025-03-26)
+
+**Note:** Version bump only for package @scaleway/configuration-loader
+
 ## [1.0.5](https://github.com/scaleway/scaleway-sdk-js/compare/@scaleway/configuration-loader@1.0.4...@scaleway/configuration-loader@1.0.5) (2024-05-15)
 
 ### Bug Fixes